+++ test/format-test.cc
@@ -292,7 +292,7 @@ TEST(WriterTest, WriteLongLong) {
   CHECK_WRITE(78ull);
   CHECK_WRITE(std::numeric_limits<long long>::min());
   CHECK_WRITE(std::numeric_limits<long long>::max());
-  CHECK_WRITE(std::numeric_limits<unsigned long long>::max());
+  CHECK_WRITE(std::numeric_limits<uint64_t>::max());
 }
 
 TEST(WriterTest, WriteDouble) {
